*In function processSkidMarks() at line 476 of file KartDriver.cxx: -Changed the skidmark rotation speed requirement from 20.0f to 110.0f. -Modified the function to work with the modified SkidMark code.
*In file SkidMark.cxx & file SkidMark.h:
-Improved skidmarks

*In file Herring.cxx:
-Commented out 3 lines which caused the program to crash at exit
if you grabbed a coin or a ? box.
-Changed a return condition in HerringInstance::update() since her shouldn't
be NULL when calling this function anymore.

*In file World.cxx:
-num_herring now is initalized to 0 at the World class constructor, since
it was just incremented and the program crashed when it was used and it was
equal to the maximum number of herrings allowed.
-In World::update(), herrings are now updated till num_herring instead of
MAX_HERRING so herrings that aren't initalized for the track aren't called
to be updated, which would cause crashes from time to time.



git-svn-id: svn+ssh://svn.code.sf.net/p/supertuxkart/code/trunk@518 178a84e3-b1eb-0310-8ba1-8eac791a3b58