Improved English grammar and punctuation
Some minor fixes in the English grammar and punctuation.
For example, in English, there shouldn’t be a space before
a colon (e.g., it should be ‘Password:’, not ‘Password :’)
(This is different than in French, where there *should*
be a space.) The commit also includes some punctuation
changes to make similar strings more consistent (e.g., trailing
periods).