Add context for key names translations.
At least in some languages most of key names is non-translatable. But some of these names are used (or can be used) in GUI in different places. For example "Play", "Pause", "End" etc...
At this moment probably the only one key name which is used in other places is "Help". But we will be safe in future.