Fix jumbled up human names in gp ranking
In the ranking shown inbetween races during a gp the names of human players have been determined
using the assertion that the human players start from the rear of the field. This is not true for
gps, where the start position is determined by the overall standings, which lead to the human's
name being displayed with the wrong kart in the rankings. This commit fixes the issue by accuratly
determining the start position of a human's kart.