Fixed mingw compilation.
It should be possible to simplify it with something like
find_library(CURL_LIBRARIES NAMES libcurl libcurldll PATHS "${PROJECT_SOURCE_DIR}/dependencies/lib")
so you can look if it works on Visual Studio.

And as far as I see pkgconfig is used only on unix. Btw. I'm also not sure if it should be if(UNIX AND NOT APPLE) ?