Allow to use dependencies for vs and mingw without changing the names every time.
Now there are:
- dependencies-vs
- dependencies-vs-64bit
- dependencies-mingw
- dependencies-mingw64

And if it's not found, then fallback to:
- dependencies
- dependencies-64bit

So that if you use only one compiler, then you can just still use "dependencies" or "dependencies-64bit".

And I didn't restrict it to WIN32 only, because in theory it should be possible to put there openglrecorder for linux build.