Add a few items, clearly separate track/modelling changes
Also, each new track get one line of changelog. Putting two on the same line vastly underestimate the importance of such a change. This also make it easier to mention removed/replaced tracks.

The changelog should not be shortened at all costs ; rather, it should be reasonably complete and not shy from using more lines as a visual cue when appropriate.