New help menu (#3322)
* Add vertical tabs to ribbon widget types

* Update GUI engine for vertical tabs

* New icons for the help menu

The slipstream icon is not great, it's here to have something now. It should be replaced by something better looking and more in STK's visual style.

* Updated help GUI

The git changelog is a bit misleading here, because the order of the items in the help menu has been revised too.

1)General (same)
2)Game modes (before : 3)
3)Powerups (before : 2)
4)Bananas (before : 5)
5)Story Mode
6)Kart classes
7)Multiplayer

* Update help screen code to handle the additional tabs

The hugely repetitive structure of one cpp/hpp per tab is kept here.

* Optimized version of new icons

Except for the slipstreaming one, as it should be replaced by a better one anyway.