Scale the weights to have 1000 as sum rather than 200
The previous 200 was meant to also limit the size of the list of item to pick among. This is not a concern anymore.

However, the interpolation for rare items (typically : switch, basketball, parachute with many karts) was quite poor as there is no good way to average 2 and 3 or 3 and 4 with integers. The very values used for the rare items were also limited because of it, though they are not fine-tuned here, under the assumption that larger testing will better help to narrow down further improvements.

And for all items, it had bigger rounding errors than now.

1000 of sum remains quite readable in the powerup.xml ; as say 235 means 23,5%.