RFC: Use Cartoon skin by default
Many users seem to prefer it as it has a much more consistent icon set.

This labels the older skins as "Classic" and sets the default skin to Cartoon.  Additionally, it changes the entry name in user_config so that existing users will have their skin setting reset on upgrade of STK (they can easily change it back if they prefer the Classic skins).