Increase default max FPS to suit 144 Hz and 165 Hz displays
The frame limiter's timer lacks precision, which seems to limit
it to millisecond timing (rather than microsecond timing).
This means we can't limit to 145 FPS (for 144 Hz displays),
as this requires a sleep duration of roughly 6.9 milliseconds
(and not 7 milliseconds).

We can however set the sleep duration to 6 milliseconds, which is good
for both 144 Hz and 165 Hz displays. Both refresh rates are becoming
increasingly common in displays used for gaming.