Greatly improve the kart-on-kart collision code
- Greatly expand the caps on mass and speed for the collision impulse factor - this makes the difference between kart classes, and between fast and slow moving karts, meaningful.
- Significantly increase the strength of kart-on-kart collision impulses to make them meaningful.
- Increase the duration of the impulse, but make it decay in strength. Instead of the kart lateral movement stopping abruptly, it is progressive, which makes for much better gameplay and appearance.
- Throw out the old code for ramming, that was completely useless, and replace it by an effective solution to assess if (1) one kart is ramming the other (2) how significant the ramming is
- Remove the squaring of the factors that was done before to (poorly) increase impulse strength

All these changes make it easier to push another kart (especially with a heavier kart against a lighter kart), but they also eliminate the dreaded collision deadlock (where one kart would push another into a rotation, and the other kart trying to not rotate away from the track direction would push back against the first kart and lock both karts until they crashed off-track).

The values of the impulse strength and of how important the ramming-factor can be are provisional, pending multiplayer playtesting, as the AI is not good to test this with.

A kart can only be subject to one impulse at a time, but for now this limitation is minor.

A more significant issue is that the impulse direction is relative to the kart. This means that if the kart is subject to a rotation while the impulse is active, the impulse will push the kart in an incorrect direction. This is very noticeable in races with many AI karts or if a kart is in an already incorrect position when pushed.