Update cake.cpp
When shooting backwards WHILE moving backwards the cake speed should be high enough to hit the target. We achieve that by considering the absolute value of  (m_owner->getSpeed()). Otherwise the speed would be negative and changed to 1.0f which is very low, and it would miss the target very often.