Tweak time-full-steer
Although discontinuous changes are maintained, there are two points of discontinuity instead of one. This allows to get an effect similar to the old time-reset-steer variable, where turning back towards a neutral straight direction is swifter.

The main goal of this patch is to improve the handling feeling of heavies (and to a lesser degree, mediums), making them more fun and slightly easier to play, without changing much their peak potential.