Reduce nitro consumption in lower difficulties
Lower difficulties collect the same number of nitro bottles in a lap as higher difficulties, but the lap lasts longer, so the proportion of time with active nitro is significantly reduced there.

This change slightly speeds up lower difficulties and increase fun from nitro. As the speed boost from nitro is smaller compared to higher difficulties (unlike in 1.X), this is not too strong.