Add an UI to edit favorite tracks
The UI works simply by ticking a checkbox in the track and GP screen and then selecting a track icon to switch its favorite status. The checkbox is always unticked by default when entering the screen. Changes are saved in the current player profile.

The display of tracks inside groups is not automatically refreshed.