Add a separate Camera tab
The User Interface tab was crowded, display didn't fit very well either, and a new tab allows to lay out camera options more directly.

The graphics settings icon is now used for the camera tab. The cartoon theme has a new icon for the graphics tab, the standard theme is using the old icon as a placeholder.