Transfer some skidding bonus power into base engine power
Before this change, skidding did not only help improve max speed, it made an enormous difference in the time needed to accelerate to max speed. Even with this change, having an active skid or completing a skid is clearly helpful, but the difference isn't as dramatic.

Engine power from skidding doesn't benefit from gear boosts, while base engine power does. Therefore, while the power is exactly equivalent with a skid boost active at the highest gear (which has a ratio of 1), it increases even beyond what an active skid would have provided in lower gears.

Therefore, to mostly compensate this effect, gear ratios for kart classes have been revised, keeping the effective engine boost close to the value substracted

An exception is seen in the first gear, where the light multiplier sees an extra reduction and the heavy multiplier is unchanged. This has little effect on acceleration performance, but addresses #5469, although the edge for heavier karts is small.