Add a parameter to adjust speed for turn radius computations
For example, with a speed-factor of 1.1 and a kart speed of 20, the game will apply the turn radius that normally corresponds to a speed of 22.

This is not meant to be used for kart classes or at the reference difficulty. Rather, this is meant to be used for lower difficulties and for handicaps, where the turn radius disadvantage of heavier kart classes tend to become much less relevant.

Also:
- Allow to adjust the value of the X elements of an InterpolationArray by a given factor
- Remove leftover references to time-reset-steer