Update the restitution array
While the previous update already made the restitution curves smoother, it assumed a quadratic relation between speed and restituted energy, whereas measurements in-game fit a nearly-linear relationship. As a consequence, an increase in kart speed could still cause a measurable decrease in post-rebound speed, albeit less than in 1.x.

Compared to the replaced values, the new values provide for a moderate increase of rebound values for medium and moderately high speed (15 to 40). The rebound force is lesser than in 1.x above speeds of about 32 and settles at 20% less at speeds of 45 or higher.

Note also that when rebounding against an obstacle without leaving the ground, the 'parking' brake apply below the parking threshold and can mostly mask the rebound. With the reduced parking threshold of 3 in the new balance code, this happens less.

The testing to validate behavior has been done with head-on collisions, as these are by far the easiest to do somewhat consistently.