Prevent unlimited speeds when the kart is falling (Fix #5237)
When falling down in the air, karts would accumulate vertical speed over time without any limit. In most jumps it wouldn't matter, but in really big jumps this could lead to extreme falling speeds and cause physics issues such as karts going through the ground.

This commit:
- Moves the base air friction formulas to btKart
- Computes the kart's vertical movement based on the direction of gravity.
- Computes air friction for the vertical component of a flying kart's movement. We purposefully only use the vertical component so that the forward movement of the kart is more or less unaffected: losing a lot of speed when jumping would make jumps much less fun.
- Applies an impulse to the kart to apply the computed force.
- Also silences an unrelated warning.

Because the friction force first slows down the kart's vertical ascent before slowing down its fall, and because the forces involved are small at low vertical speeds, the effect on small and medium jumps is very limited. The effect is most noticeable when the jump is big and when the heights of the take-off point and of the landing point are significantly different.

As the same force is currently applied to all kart classes, the terminal vertical velocity of a kart depends on its class (~31 for lights, ~36 for medium, ~41 for heavies). This could be changed if needed, but it's probably fine as is.

This commit has been tested in various situations, including huge jumps (about 3.5s falling from a neutral vertical speed, leading to a falling speed of over 60 before this commit) and situations with gravity textures and changing gravity (flying around Interplanetary's moon).