Use the main kart version to know the animations format
Instead of adding a new XML node, it's simpler to rely on the existing kart version scheme. This means that karts exported for Evolution won't be readable by 1.x binaries without manual modification, but that's probably not worse than all the renamed animations not working and the cause of the issue not being clear.

The specifications for version 4 are not yet final: the new and renamed animations are part of it, but other elements may be changed going forward if it makes sense.

Also:
- Add some comments to roughly explain what each supported kart format version corresponds to.
- Avoid reporting the placeholder as a missing animation in the artist-debug-mode logging.