Change how mass influences kart collisions
Previously, a kart that was X times heavier than another would receive an impulsion 1/X times stronger while the other kart would receive an impulsion X times stronger, leading to a high compounded differences. For example, a 40% disparity in mass would result for example in an almost 100% disparity in impulse strength. It now results in a 65% disparity in impulse strength.

This commit also makes related adjustments:
- As the peak impulse strength (with a heavy hitting a light kart) has been reduced by 13%, the base strength of collisions is increased by ~8%.
- The peak bonus from speed difference is obtained with a differential of 21 or more, instead of 18 or more. This reduces somewhat the capacity of a lighter kart with some extra speed to push off a heavier kart.