Adjust drifting bonuses for lower difficulties
In lower difficulties, it becomes much easier to get the stronger boost levels, as the time required to get a boost level are the same but slower karts mean more time spent in a given curve. Therefore, the speed advantage of 2nd and 3rd tier drifts are made less significant.

Those changes also reduce the advantage of good drifting usage in low difficulties, and make the completion of a 3 laps race in Casual less likely to be faster than a 4 laps SuperTux race. However, it's still not quite enough for this latter purpose.

This commit also fixes the max-speed-reverse-ratio for the Novice difficulty, which was erroneously increased instead of decreased when the base max speed was increased.