Move the graphical presets out of the video options screen
They are defined in their own namespace and the initialization phase is run during the startup phase of the game.

The video option screen simply retrieves the relevant variable and works much as before.

However, this will allow much easier access to the presets in other contexts, as creating an OptionsScreenVideo object is not necessary anymore:
- To easily set up a given graphical preset from the command line
- To offer recommandations through the video settings suggestion tool. This was only partially implemented before 1.5 and hidden in the UI, but this commits removes one of the main barrier that remained to get it done.