Update the nitro efficiency display
We now account for potential variations in the minimum burst duration, multiplying by the ratio of minimum boost duration to minimum duration with consumption. This neglects the fade-out-time, but it's good enough for what is just a general guidance for players.

This commit also adds an extra variable to avoid calling getCombinedCharacteristic repeatedly (mainly to make the code more readable).