Minor code quality improvement in ChatCommands
I had initially considered having the extra command fall-through to the main command to display both messages and remove the need for the extra command call, but while it made code denser it also made it less readable.

Instead, we simply use return in the switch.

The 'extra' messages are also placed first as they describe the purpose of the command and that's better done before the explanations on how to use the command.